Conflicts:
	app/controllers/TaskCommentsController.php
	app/controllers/TaskController.php
	app/helpers.php
	app/models/Opportunity.php
	app/models/Task.php
	app/routes.php
	app/views/crmcomments/comments.blade.php
	app/views/opportunityboards/board.blade.php
	app/views/opportunityboards/index.blade.php
	app/views/opportunityboards/manage.blade.php
	app/views/taskboards/board.blade.php
	app/views/taskboards/manage.blade.php
	app/views/taskboards/taskmodal.blade.php
	dbv-1/data/schema/prc_GetTasksBoard.sql
	dbv-1/data/schema/prc_GetTasksGrid.sql
	dbv-1/data/schema/tblOpportunityComments.sql
	dbv-1/data/schema/tblTask.sql

+++ app/controllers/RateTablesController.php
@@ -4,10 +4,15 @@ class RateTablesController extends \BaseController {
 
     public function ajax_datagrid() {
         $CompanyID = User::get_companyID();
-        $rate_tables = RateTable::join('tblCurrency', 'tblCurrency.CurrencyId', '=', 'tblRateTable.CurrencyId')->where(["tblRateTable.CompanyId" => $CompanyID])->select(["tblRateTable.RateTableName","Code","tblRateTable.updated_at", "tblRateTable.RateTableId"]);
+        $rate_tables = RateTable::
+        join('tblCurrency','tblCurrency.CurrencyId','=','tblRateTable.CurrencyId')
+            ->join('tblCodeDeck','tblCodeDeck.CodeDeckId','=','tblRateTable.CodeDeckId')
+            ->select(['tblRateTable.RateTableName','tblCurrency.Code','tblCodeDeck.CodeDeckName','tblRateTable.updated_at','tblRateTable.RateTableId'])
+            ->where("tblRateTable.CompanyId",$CompanyID);
+        //$rate_tables = RateTable::join('tblCurrency', 'tblCurrency.CurrencyId', '=', 'tblRateTable.CurrencyId')->where(["tblRateTable.CompanyId" => $CompanyID])->select(["tblRateTable.RateTableName","Code","tblRateTable.updated_at", "tblRateTable.RateTableId"]);
         $data = Input::all();
         if($data['TrunkID']){
-            $rate_tables->where('TrunkID',$data['TrunkID']);
+            $rate_tables->where('tblRateTable.TrunkID',$data['TrunkID']);
         }
         return Datatables::of($rate_tables)->make();
     }
@@ -110,14 +115,15 @@ class RateTablesController extends \BaseController {
      * @return Response
      */
     public function view($id) {
-            $trunkID = RateTable::where(["RateTableId" => $id])->pluck('TrunkID');
-            $countries = Country::getCountryDropdownIDList();
-            $CodeDeckId = RateTable::getCodeDeckId($id);
-            $CompanyID = User::get_companyID();
-            $codes = CodeDeck::getCodeDropdownList($CodeDeckId,$CompanyID);
-            $isBandTable = RateTable::checkRateTableBand($id);
-            $code = RateTable::getCurrencyCode($id);
-            return View::make('ratetables.edit', compact('id', 'countries','trunkID','codes','isBandTable','code'));
+        $rateTable = RateTable::find($id);
+        $trunkID = RateTable::where(["RateTableId" => $id])->pluck('TrunkID');
+        $countries = Country::getCountryDropdownIDList();
+        $CodeDeckId = RateTable::getCodeDeckId($id);
+        $CompanyID = User::get_companyID();
+        $codes = CodeDeck::getCodeDropdownList($CodeDeckId,$CompanyID);
+        $isBandTable = RateTable::checkRateTableBand($id);
+        $code = RateTable::getCurrencyCode($id);
+        return View::make('ratetables.edit', compact('id', 'countries','trunkID','codes','isBandTable','code','rateTable'));
     }
 
 
@@ -128,13 +134,31 @@ class RateTablesController extends \BaseController {
                             ->where("tblRateTable.RateTableId", $id)->count();
             //Is RateTable assigne to RateTableRate table then dont delete
             if ($is_id_assigned == 0) {
-                if (RateTable::where(["RateTableId" => $id])->delete()) {
-                    return Response::json(array("status" => "success", "message" => "RateTable Successfully Deleted"));
-                } else {
-                    return Response::json(array("status" => "failed", "message" => "Problem Deleting RateTable."));
+                if(!empty(RateTable::checkRateTableInCronjob($id))){
+                    if(RateTableRate::where(["RateTableId" => $id])->count()>0){
+                        if (RateTableRate::where(["RateTableId" => $id])->delete() && RateTable::where(["RateTableId" => $id])->delete()) {
+                            return Response::json(array("status" => "success", "message" => "RateTable Successfully Deleted"));
+                        } else {
+                            return Response::json(array("status" => "failed", "message" => "Problem Deleting RateTable."));
+                        }
+                    }else{
+                        if (RateTable::where(["RateTableId" => $id])->delete()) {
+                            return Response::json(array("status" => "success", "message" => "RateTable Successfully Deleted"));
+                        } else {
+                            return Response::json(array("status" => "failed", "message" => "Problem Deleting RateTable."));
+                        }
+                    }
+
+                }else{
+                    return Response::json(array("status" => "failed", "message" => "RateTable can not be deleted, Its assigned to CronJob."));
                 }
+
             } else {
-                return Response::json(array("status" => "failed", "message" => "RateTable can not be deleted, Its assigned to Customer Rate."));
+                if(empty(RateTable::checkRateTableInCronjob($id))){
+                    return Response::json(array("status" => "failed", "message" => "RateTable can not be deleted, Its assigned to Customer Rate and CronJob."));
+                }else{
+                    return Response::json(array("status" => "failed", "message" => "RateTable can not be deleted, Its assigned to Customer Rate."));
+                }
             }
         }
     }
@@ -283,21 +307,32 @@ class RateTablesController extends \BaseController {
     
     public function exports($type) {
             $CompanyID = User::get_companyID();
-            $rate_tables = RateTable::where(["CompanyId" => $CompanyID])->orderBy("RateTableId", "desc");
+            /*$rate_tables = RateTable::where(["CompanyId" => $CompanyID])->orderBy("RateTableId", "desc");
             $data = Input::all();
             if($data['TrunkID']){
                 $rate_tables->where('TrunkID',$data['TrunkID']);
             }
-            $rate_tables = $rate_tables->get(["RateTableName"]);
-
+            $rate_tables = $rate_tables->get(["RateTableName"]);*/
+            $rate_tables = RateTable::
+            join('tblCurrency','tblCurrency.CurrencyId','=','tblRateTable.CurrencyId')
+                ->join('tblCodeDeck','tblCodeDeck.CodeDeckId','=','tblRateTable.CodeDeckId')
+                ->select(['tblRateTable.RateTableName','tblCurrency.Code as Currency Code','tblCodeDeck.CodeDeckName'])
+                ->where("tblRateTable.CompanyId",$CompanyID);
+            //$rate_tables = RateTable::join('tblCurrency', 'tblCurrency.CurrencyId', '=', 'tblRateTable.CurrencyId')->where(["tblRateTable.CompanyId" => $CompanyID])->select(["tblRateTable.RateTableName","Code","tblRateTable.updated_at", "tblRateTable.RateTableId"]);
+            $data = Input::all();
+            if($data['TrunkID']){
+                $rate_tables = $rate_tables->where('tblRateTable.TrunkID',$data['TrunkID']);
+            }
+            $rate_tables = $rate_tables->get();
             $excel_data = json_decode(json_encode($rate_tables),true);
+
 
             if($type=='csv'){
                 $file_path = getenv('UPLOAD_PATH') .'/Rates Table.csv';
                 $NeonExcel = new NeonExcelIO($file_path);
                 $NeonExcel->download_csv($excel_data);
             }elseif($type=='xlsx'){
-                $file_path = getenv('UPLOAD_PATH') .'/Rates Table.xlsx';
+                $file_path = getenv('UPLOAD_PATH') .'/Rates Table.xls';
                 $NeonExcel = new NeonExcelIO($file_path);
                 $NeonExcel->download_excel($excel_data);
             }
@@ -326,7 +361,7 @@ class RateTablesController extends \BaseController {
                 $NeonExcel = new NeonExcelIO($file_path);
                 $NeonExcel->download_csv($rate_table_rates);
             }elseif($type=='xlsx'){
-                $file_path = getenv('UPLOAD_PATH') .'/'.$RateTableName . ' - Rates Table Customer Rates.xlsx';
+                $file_path = getenv('UPLOAD_PATH') .'/'.$RateTableName . ' - Rates Table Customer Rates.xls';
                 $NeonExcel = new NeonExcelIO($file_path);
                 $NeonExcel->download_excel($rate_table_rates);
             }
@@ -345,6 +380,7 @@ class RateTablesController extends \BaseController {
         $RateTableRate['Rate'] = $data['Rate'];
         $RateTableRate['Interval1'] = $data['Interval1'];
         $RateTableRate['IntervalN'] = $data['IntervalN'];
+        $RateTableRate['ConnectionFee'] = $data['ConnectionFee'];
         $rules = RateTableRate::$rules;
         $rules['RateID'] = 'required|unique:tblRateTableRate,RateID,NULL,RateTableId,RateTableId,'.$id.',EffectiveDate,'.$data['EffectiveDate'];
         $validator = Validator::make($RateTableRate, $rules);
